+++ examples/mem_trace/src/lib.rs
@@ -107,7 +107,7 @@ impl Args {
 }
 
 // 1 MiB = 2**20
-const CHANNEL_SIZE: usize = 1 << 20;
+const CHANNEL_SIZE: u32 = 1 << 20;
 
 struct Instrumentor<'c> {
     ctx: Mutex<nvbit_rs::Context<'c>>,
